refactor(brutus): extract LLM utilities to llm.go
